AX: Add debug IPC for web process bits, limited tree dump https://bugs.webkit.org/show_bug.cgi?id=295747 rdar://155563271
Reviewed by NOBODY (OOPS!).

This patch adds some new IPC messages to provide better accessibility debug
tooling down the road. These additions and enhancements include:
- New IPC for if accessibility is enabled in the web process
- New IPC for whether the accessibility thread has been initialized
in the web process
- Updating the AXObjectCache::treeData() method to not return the renderer,
node, or outerHTML info by default. This allows for easier comparison between
the live and isolated trees. To not regress existing calls to this method, I
added in those stream options to the other call sites.

* Source/WebCore/accessibility/AXLogger.cpp:
(WebCore::operator<<):
(WebCore::streamAXCoreObject):
* Source/WebCore/accessibility/AXLogger.h:
* Source/WebCore/accessibility/AXObjectCache.cpp:
(WebCore::AXObjectCache::treeData):
* Source/WebCore/accessibility/AXObjectCache.h:
* Source/WebCore/accessibility/isolatedtree/AXIsolatedTree.h:
* Source/WebCore/accessibility/mac/AXObjectCacheMac.mm:
(WebCore::AXObjectCache::initializeAXThreadIfNeeded):
(WebCore::AXObjectCache::isAXThreadInitialized):
* Source/WebCore/accessibility/mac/WebAccessibilityObjectWrapperMac.mm:
(-[WebAccessibilityObjectWrapper _accessibilityPrintTrees]):
* Source/WebCore/page/Page.cpp:
(WebCore::Page::accessibilityTreeData const):
* Source/WebCore/page/Page.h:
* Source/WebKit/UIProcess/API/C/mac/WKPagePrivateMac.h:
* Source/WebKit/UIProcess/API/C/mac/WKPagePrivateMac.mm:
(WKPageAccessibilityIsEnabled):
(WKPageAccesssibilityThreadInitialized):
* Source/WebKit/UIProcess/API/mac/WKWebViewPrivateForTestingMac.h:
* Source/WebKit/UIProcess/API/mac/WKWebViewTestingMac.mm:
(-[WKWebView _retrieveAccessibilityTreeData:]):
(-[WKWebView _retrieveLimitedAccessibilityTreeData:]):
* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::isAccessibilityEnabled):
(WebKit::WebPageProxy::isAccessibilityThreadInitialized):
(WebKit::WebPageProxy::getAccessibilityTreeData):
* Source/WebKit/UIProcess/WebPageProxy.h:
* Source/WebKit/WebProcess/WebPage/WebPage.cpp:
(WebKit::WebPage::isAccessibilityEnabled):
(WebKit::WebPage::isAccessibilityThreadInitialized):
(WebKit::WebPage::getAccessibilityTreeData):
* Source/WebKit/WebProcess/WebPage/WebPage.h:
* Source/WebKit/WebProcess/WebPage/WebPage.messages.in:

+++ Source/WebCore/accessibility/isolatedtree/AXIsolatedTree.h
@@ -53,7 +53,7 @@ class AXGeometryManager;
 class AXObjectCache;
 class AccessibilityObject;
 class Page;
-enum class AXStreamOptions : uint8_t;
+enum class AXStreamOptions : uint16_t;
 
 static constexpr uint16_t lastPropertyFlagIndex = 21;
 // The most common boolean properties are stored in a bitfield rather than in a HashMap.
